backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.7.0
google-generativeai>=0.5.0
requests>=2.31.0
pdfplumber>=0.11.0
PyPDF2>=3.0.0
pypdf>=4.0.0
python-docx>=1.1.0
reportlab>=4.2.0
httpx>=0.27.0
aiofiles>=23.2.0
